[advance-reboot] Create summarized report of reboot timing data (#3453)
Create a summarized version of advanced-reboot timing data.
The timestamps in the detailed version of reboot-timing-data aren't necessarily useful in data plotting. Also not all the entities have Starting/Stopping data. This leads to inconsistent data being uploaded, and some columns have no values.

Create a summary (along with a detailed report) version of the advanced-reboot timing data.
The summarized version contains the time-taken data, and not the timestamps. This is useful in making kusto table and queries simpler.
The detailed version will still be generated and can be used for further analysis.