[sanity_check]: Change order of mux stop on ToRs (#3479)
An upcoming image change will cause the switchover performed during the sanity check to be reversed (since stopping the mux container will trigger a switchover). Since the upper ToR mux container is stopped last, a switchover to make the lower ToR active will be initiated, which has the potential to conflict with with switchover to the upper ToR on the next line. Stopping the upper ToR first resolves this.

Signed-off-by: Lawrence Lee <lawlee@microsoft.com>