Fix the mux active/standby change issue for dualtor FIB testing
On dualtor testbed, the active/standby status may be changed while the PTF script is running.
Consequently the PTF script testing FIB may fail.

The reason of active/standby change is that no ICMP responder is running in PTF to simulate
servers. Then status of the muxes are unhealthy and link manager may try to recover the muxes
by switching the active/standby side.

The fix is to start ICMP responder befor the FIB testing to avoid mux active/standby state flap.

Another logic issue of the FIB testing is getting active side of the DUT index. The issue is fixed
by changing:
    if not target_dut_index:
to:
    if target_dut_index is None:
Initializing `target_mac` also has the similiar issue and is fixed in the same way.

The third fix is to change `ptf_test_port_map` and `get_mux_status` from fixture
to function. These two functions depends on fixtures like `set_mux_random`
and `set_mux_same_side`. The execution sequence matters. If they are all
fixtures, changing the sequence of using the fixtures in test function argument
list may affect the test logic and cause issues that are hard to debug.
Making two of them to function calls can avoid this potential volunerability.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>