Fix test_acl on dualtor testbed (#3507)
Description of PR
Summary:

This PR is to fix test_acl on dualtor testbed.

test_acl failed on dualtor testbed because 2 reasons

The mac address of Vlan interface and L2 interfaces are different. So we should use different MAC address for upstream packets and downstream packets. For upstream packets, the MAC address of vlan interface should be used. For downstream packets, L2 interface's MAC address should be used.
The DUT that are selected for testing should be active.
Signed-off-by: bingwang bingwang@microsoft.com

Approach
What is the motivation for this PR?
This PR is to fix test_acl on dualtor testbed.

How did you do it?
Toggle mux cable to randomly selected DUT;
Use different MAC address for upstream packets and downstream packets.
How did you verify/test it?
Verified on both dualtor/single tor testbed. All passed.

Signed-off-by: bingwang <bingwang@microsoft.com>