Fix test_vxlan_decap on dualtor testbed (#3508)
Description of PR
Summary:
This PR is to fix test_vxlan_decap on dualtor testbed.

There are two reasons for the failure of test_vxlan_decap on dualtor testbed.

The MAC address of vlan interface and L2 interfaces are different. We should use correct MAC address for generated packets
The DUT selected for testing may not be the active one. A toggle is necessary to ensure that the test are running on active ToR
Signed-off-by: bingwang bingwang@microsoft.com

Approach
What is the motivation for this PR?
This PR is to fix test_vxlan_decap on dualtor testbed.

How did you do it?
Toggle mux to selected ToR before testing
Use correct MAC address for generated packets.
How did you verify/test it?
Verified on both T0 and dualtor testbeds. All passed.

Signed-off-by: bingwang <bingwang@microsoft.com>