[platform_tests][api] Add hardware revision test case to chassis and psu (#3466)
Description of PR
Adds tests cases to validate the existence of hardware revision for chassis and psu devices.

Approach
What is the motivation for this PR?
We recently defined the hardware revision API and implemented it on Mellanox platforms. We now desire to add testing to verify that this platform API implementation remains stable so it may be relied upon in other areas of code.

How did you do it?
Used the same method as in other api tests such as get_serial() and get_model() for both devices.