[qos] enable QoS test on dualtor topology (#3513)
What is the motivation for this PR?
qos test is failing on dualtor topology. It was enabled for dualtor-56 but not for dualtor.

How did you do it?
Add 'dualtor' to supported list.
For topologies that are not on the supported list. Skip the test instead of failing.

How did you verify/test it?
Tested the skip change first. Making sure the tests are skipping instead of failing.
Then added 'dualtor' to run tests and run into passes and legit failures.

Signed-off-by: Ying Xie <ying.xie@microsoft.com>