[minigraph]: Add the ability to set a per-port speed in the minigraph
In the case of devices with mixed port speeds, the speed for each port
has to be individually specified in the minigraph, and a single speed
for the entire device cannot be used. Instead, read the connection
graph, which contains the speed for each individual link, and use the
speed specified there for each port.

If a port/link isn't specified in the connection graph, then the speed
for each port in port_config.ini is still used. If no speed is specified
there, then the speed in the inventory yaml file is used.

Signed-off-by: Saikrishna Arcot <saiarcot895@gmail.com>