[dualtor orchagent] Fix failures in case standby_mux_toggle_active (#3529)
Fix for Test test_downstream_standby_mux_toggle_active. The failure is due to:
Wrong destination address for the route added.
Wrong tor selected for monitoring traffic packets.
How did you do it?
Add a route to a address in 10.x network. Earlier this route was added to the server which also was the nexthop address.
Fix the test by using correct torhost objects to monitor server and tunnel traffic respectively.