Enable ansible pipelining to allow modules to execute more than one cmds (#3530)
Tests using server_traffic_utils are flaky on real dualtor testbed. The step in which tcpdump is started on vmhost fails.
It turns out that this inconsistent error is due to Ansible limiting number of SSH operations that can be performed:
When disabled, the server_traffic_utils can start tcpdump on vmhost as expected.