minigraph: Rework the connection graph data usage
Instead of getting and using the connection graph data in port_alias.py,
use it instead in the template file, and add a new ansible task to load
in the connection graph data. This keeps port_alias.py strictly for
loading in the port_config.ini data, and instead moves the speed
selection logic into the template file.

Signed-off-by: Saikrishna Arcot <saiarcot895@gmail.com>