minigraph: Add the ability to set a per-port speed in the minigraph (#3527)
* minigraph: Add the ability to set a per-port speed in the minigraph

In the case of devices with mixed port speeds, the speed for each port
has to be individually specified in the minigraph, and a single speed
for the entire device cannot be used. Instead, read the connection
graph, which contains the speed for each individual link, and use the
speed specified there for each port.

The speed specified in the connection graph is used only if
port_config.ini doesn't have a speed specified there. If there is a speed
specified there, then that is used.

If no speed is specified in port_config.ini, and there's no speed specified
for the port in the connection graph, then the speed in the inventory
yaml file is used.

Signed-off-by: Saikrishna Arcot <saiarcot895@gmail.com>