Add space before extra arguments for renumber_topo and restart_ptf
In the testbed-cli.sh tool, extra arguments can be passed to the playbook for
renumber_topo and restart_ptf. However, there is no whitespace before the
extra arguments `$@`. Then the extra arguments will be part of value of ansible
variable `ptf_ipv6`. This command works:
  ./testbed-cli.sh restart-ptf <testbed_name> password.txt

  This command will fail:
  ./testbed-cli.sh restart-ptf <testbed_name> password.txt -vv

The last "-vv" will be appended to value of ansible variable `ptf_ipv6`.

The fix is just to add a white space before `$@`.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>