[topo] Add test topology for 7050QX-32S-S4Q31 (#3568)
Signed-off-by: Saikrishna Arcot <saiarcot895@gmail.com>

Adds a topology definition for 31 40G ports and 4 10G/1G ports, where those 10G ports are not used for automated tests.

What is the motivation for this PR?
This is for use with Arista 7050QX-32S-S4Q31.

How did you verify/test it?
Deployed a minigraph with this topology and checked to make sure that the VLAN assignments and the port channel assignments were correct. No automated tests were run.