[dut utils] protect against empty dut vars (#3581)
What is the motivation for this PR?
When running test against a none existing dut, e.g. a wrong inventory was chosen, dut_util will throw exception due to host_vars are None, which is not directly helping finding the issue.

How did you do it?
Protect against host_var being none or empty.

How did you verify/test it?
Run a test with a wrong inventory specified. With the change. there is still an exception, but this time the exception is clear that the host name is not found.

Signed-off-by: Ying Xie <ying.xie@microsoft.com>