[show_interface] Check for namespace when internal intfs need to be displayed  (#3630)
The current code assumes that namespace is always present when the include_internal_intfs option is enabled and the cmd is constructed to display the interface status. Failure was seen when trying to run pytest cases against 201811 branch. In one of the tests, include_internal_intfs was set to True and module run failed because the cmd was unsupported in 201811 branch. To avoid that failure, check for presence of namespace as well along with the current flag to extend the interface cmd options

Signed-off-by: Neetha John <nejo@microsoft.com>

How did you verify/test it?
Ran the failing tests on 201811 and it passed