[rdma] Enable minigraph generation for breakout 7050 (#3632)
In the current tgen topology, the connection graph includes only the links connected to the traffic generator. Hence the 10G links are not defined. The current logic for defining the speed for interfaces in the minigraph follows the order below
1. get the speed from port_config.ini.
2. if not obtained from 1, then get it from connection graph
3. if not obtained from 2, then set the value of 'iface_speed'
   
We will be setting the speed incorrectly to 40G (condition 3 kicks in) for the 10G links because no speed is defined in port_config.ini for this SKU and 2) is also not satisfied for the reason stated above.
To overcome this problem, added a new file where link-speed mappings can be defined for each sku, if needed and use that info as well in the minigraph generation to define the correct speed for a link.

Signed-off-by: Neetha John <nejo@microsoft.com>

How did you verify/test it?
Successfully generated minigraph for the 7050 breakout sku with the changes and ensured that the speeds for the links were set as desired.