Check route redistribution for static routes (#3633)
What is the motivation for this PR?
The static routes applied to CONFIG_DB are supposed to get redistributed to the BGP neighbors. The motivation of the PR is to check if the static routes get properly redistributed.

How did you do it?
Check route advertised to BGP neighbor after configuring static routes.