[advanced-reboot] Fix for "SSH threads haven’t finished for 300 seconds"
This is due to "quit" command received but v4_routing_ok or
v6_routing_ok is not true, than we will wait forever in
self.queue.get(). What we need to do is to retry this loop without
fetching command (we are quiting, no more commands are expected) until
v4_routing_ok and v6_routing_ok are true or timeout will be reached.

Signed-off-by: Stepan Blyschak <stepanb@nvidia.com>