[qos] Add qos params for breakout 7050 and enable test run (#3658)
Arista-7050-QX-32S has 4 10G ports and 31 40G ports. Shared space is different from the 32 port SKU and new parameters need to be added for the qos test run. The testcases have also been modified to consume this information

Signed-off-by: Neetha John <nejo@microsoft.com>

How did you verify/test it?
Ran the qos sai test with the new values on breakout 7050 and it passed