Improve advanced reboot test to support multiple VLANs (#3650)
What is the motivation for this PR?
The current advanced reboot tests only support the scenario with one vlan and no portchannel in vlans. This PR aims to improve the test cases to cover broader scenarios.

How did you do it?
Update advanced reboot to take account of multiple VLANs and ignore portchannel ports in VLANs.