[pytest] Test the feature of memory checker. (#3582)
What is the motivation for this PR?
This PR aims to test the feature of memory checker in 20191130.73 and newer image version. Currently this pytest script only tests the feature on streaming telemetry container and will extend this test to all other containers once this feature is fully implemented.

How did you do it?
I implemented this test according to the following steps:
Installs stress utility in container under testing.
Invokes stress utility to eat up the amount of memory which is larger than the threshold.
Waits for the container to be stopped and restarted.
Checks whether all critical processes in container are running.

How did you verify/test it?
I tested this feature on the testbed str-msn2700-01.