Fixing generic config reload option to use force if supported (#3698)
#### What is the motivation for this PR?
With the introduction of system checks for config reload, without the system being stable, config reload will not proceed. However many tests in the sonic-mgmt test suite use config reload to recover the system for the test cases where they intentionally kill or stop a docker or process. Hence in order to support those these and not to have any impact on the existing tests, the config_reload API will use -f option if supported.

#### How did you do it?
Added logic to do force config reload if supported