Refactor the mux simulator (#3699)
* Refactor the mux simulator

The original mux simulator needs to run ovs-vsctl command to get the flow state for each
mux status query. When the dualtor setup is running, each of the ToR queries for each of
its mux status at a short interval. This caused performance issue.

Another issue of the original mux simulator is that while switching active/standby side,
the flow needs to be removed firstly and then add back. While the flow is being recreated,
the mux simulator may receive request querying mux status. Because of the conflict, an
error may be responded for the status query. This will result in "Unknown" mux status in
DUT.

This change totally re-designed the mux-simulator. Major improvements:
* Introduced new class `Mux` represents a mux bridge.
* Introduced new class `Muxes` represents a collection of mux bridges.
* Status of each mux is stored as attribute of `Mux` instance. Getting mux status just needs
  to read the instance attributes in memory. No need to run `ovs-ofctl` command on server.
* While updating the flow of a mux bridge, all the instance attributes are updated to
  reflect the actual configuration.
* Introduced lock to ensure that reading status and updating configuration won't happen at
  the same time to ensure consistency.
* Improved log messages. The mux simulator logs would be more helpful for troubleshooting.
* Improved exception handling. The view functions are cleaner and more elegant.
* Although lock is introduced, we also set threaded to False for the flask HTTP server.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>