[test_vlan] teardown: Handle cleanup for DUT and PTF separately (#3321)
The new PTF interfaces added during the test are not deleted in teardown if the DUT is not reachable/reponding.
In the present code, DUT and PTF cleanup happens inside one try-catch block. BEcause of this, if DUT cleanup throws exception, PTF cleanup is also skipped.
Separate DUT and PTF cleanup to two different try-catch blocks. Now if DUT cleanup throws exception, PTF cleanup wll still be executed.