Skip running tests if pretest failed
When we use the run_tests.sh tool to run test, it may run some pretests
to prepare DUT. In case pretests failed, its highly possible that the
DUT is not in good status. There would be no point continue running
rest of the tests that may take many hours to complete.

This change enhanced the run_tests.sh script to skip running tests if
pretest failed.

Signed-off-by: Xin Wang <xiwng5@microsoft.com>