[crm] skip clear percentage threashold check while percentage is 0
What is the motivation for this PR?
To fix the failures in clear percentage threshold check while percentage is 0

How did you do it?
SWSS print warning log for THRESHOLD_EXCEEDED and THRESHOLD_CLEAR coupled. So if the test case need to skip check one, it also needs to skip the other.

How did you verify/test it?
Run test_crm.py, it should pass.

Signed-off-by: Kevin Wang <shengkaiwang@microsoft.com>