Changes to run QOS testcase on 7260 64x100 with t1-64-lag topology (#3713)
Following changes are done to support running QOS testcases on 7260 64x100 with t1-64-lag topology -

Added Notion of topo in qos.yml. This is needed because for same asic based on topology (t0/t1) mmu profile can be different and need to specify it's own set of parameters.

Added support to pass explicitly source/destination ptf port if from qos.yml. This is needed for some asic there is port to buffer mapping and the port generated by test cases might not adhere to that.

In API active_ip_interfaces() added ping count. This was done because it's possible as part of test case ARP entry can become stale and ping will fail but HW entry is fine. Without this change test case that passes explicit port id gets filter out since they are ok to send/receive traffic but SW ping is failing.