[dualtor][orchagent] Change the check time of queue counter
What is the motivation for this PR?
Support to check queue counters up to 60s for the timing issue.

How did you do it?
Call the wait_until function to check the queue counters every 5s continuously. The timeout is set to 60s.

How did you verify/test it?
Run the queue counters relates test cases that should be passed, such as test_tor_ecn.py.

Signed-off-by: Kevin Wang <shengkaiwang@microsoft.com>