added http copy test (#3787)
Summary:
Adds a test that checks that the DUT is able to access external objects via HTTP.

Approach
What is the motivation for this PR?
This test existed in the Arista testplan but was not present for SONiC.

How did you do it?
An HTTP server is started on the localhost where a large file is downloaded. The DUT uses curl to fetch the file from the HTTP server. File integrity is checked by comparing checksums for the file before and after transport.

How did you verify/test it?
The test passed on my platform and confirmed that the file was copied to the host and the dut manually.

Any platform specific information?
I've been encountering the issue found in #3778 , so I've had to skip the sanity checks to confirm that this test passes. This shouldn't be an issue since this test does not interact with the T1 leaf nodes in any way. The only communication necessary is between the host and the DUT.

Co-authored-by: Andoni Sanguesa <t-asanguesa@microsoft.com>