[topo] Bug fix: wrong ipv6 link local address on backplane port in t0-56 (#3793)
There are 8 VMs in t0-56 topo. The 1st half of VMs share the same IPv6 link-local addresses on backplane ports with the 2nd half, causing BGPv6 neighbour flapping.
This is to fix the issue.
We also reassigned the IP addresses in look back and back plane interfaces in the same way it has been done in t0 64 topo. By doing so those IP addresses will be easy to remember.