Fix PTF Container Down after HTTP Copy Test (#3829)
What is the motivation for this PR?
Previous version of the http copy test was crashing the PTF container. This modification fixes that issue.

How did you do it?
The error originally occurred in closing the HTTP server remotely. Instead of trying to get the pid of the http server process by communicating with the PTF, a script is run on the PTF that identifies the process and shuts it down.

How did you verify/test it?
After the test I ran a second test that would check for communication with the PTF container. Both tests passed while not shutting down the container.

Co-authored-by: Andoni Sanguesa <t-asanguesa@microsoft.com>