Merged PR 23: Use pipelineName_buildId as testbed lock/release user
Use pipelineName_buildId as testbed lock/release user

For multiple pipelines using same testbed, or for multiple run of a same pipeline,
they will try to lock a testbed that has already been locked. Currently the code
uses same user "Azure Pipeline" to lock the testbed. The competing lock attempts
are considered as initiated by same user. The testbed booking system will simply
update the existing lock with new release time. The result is that multiple pipelines
or even multiple runs of a same pipeline can work on a same testbed. This would
cause conflicts using testbeds and unnecessary nightly test failures.

The fix is to improve the lock/release script to use pipeline name + build ID as
lock/release user. Then each run would have exclusive lock of a testbed. Subsequent
runs using the same testbed can not override the previous lock.

The drawback is that in case something went wrong and a testbed is not released by
pipeline, then we have to manually release the testbed. Luckly, each lock reserves
a testbed for 36 hours. When it was not released by pipeline, the lock would eventually
timeout without manual intervention.

In case nightly test of a testbed lasts longer than 24 hours, the nightly test of next day
may fail to lock the testbed. This is fine. We can wait another day, then the nightly test
should be able to lock the testbed because the lock of the first run has been either
released or timeout.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>