[pytest] Fix the fast failure when checking the format of Monit alerting message (#3809)
What is the motivation for this PR?
This PR aims to fix the fast failure issue when checking the format of Monit alerting message. Initially check_monit_last_output(...) was called by wait_until(...) to continuously check whether the format of Monit alerting message was correct or not. If the Monit alerting message was not found, this script should return false instead of failing this test directly.

How did you do it?
I replaced the statement pytest.fail(...) in the function check_monit_last_output(...) with return False. At the same time, some logging statements are added in order to show what action was done during testing.

How did you verify/test it?
I test this change on the DuT str-msn2700-22.

Any platform specific information?
N/A

Supported testbed topology if it's a new test case?
N/A