[multi-asic][chassis] fix the port index in port_index_map for masic (#3897)
What is the motivation for this PR?
The change done in PR #3643 to use the get_port_map, does not work for multi-ASIC platforms. The existing logic in config_facts gives the same index for ports on different ASICs.

The PR is a fix for this problem. get_port_map will use the port index present in the config_db.

How did you do it?
Change in config_facts to get the port index from config_db if its available. If the index is not present use the exisiting lofic to generate the index

How did you verify/test it?
run the test platform_tests/rest_reload_config.py to verify

Signed-off-by: Arvindsrinivasan Lakshmi Narasimhan <arlakshm@microsoft.com>