[Dynamic Buffer Calc] Add test cases for dynamic buffer calculation on top of port auto-negotiation (#3647)
- What is the motivation for this PR?
Add test cases for dynamic buffer calculation on top of port auto-negotiation

- How did you do it?
Add test cases for port auto-negotiation.
Add different configurations for port auto-negotiation and verify the behavior of dynamic buffer calculation
Skip the test if speed_to_test is not supported by the switch
There is a logic to verify the calculated headroom size in the test:
to compare the calculated headroom against the predefined pg_profile_lookup.ini
Some speeds that supported by SAI, like 1G, isn't defined in the file because they are very unlikely to be used
However, we need to take that speed for testing in some testbeds. Skip this verification in that case.

- How did you verify/test it?
Run the regression test.