Add wait for critical services in health check for advanced reboot (#3915)
What is the motivation for this PR?
There is a health check of the dut recently introduced to advanced reboot tests which checks the status of services including snmp. However, snmp starts three minutes after start-up and the advanced reboot tests do not have a dependency on snmp start. It follows that there is a racing condition between the test cases check the state of snmp and snmp actually starts. If the test case checks snmp state earlier that it starts, the test would fail even though it would come up a few moments later. This PR aims to fix the false alarm.

How did you do it?
Add wait for critical services in the health check for advanced reboot tests.