Update VLAN prefixes (#3898)
What is the motivation for this PR?
The two-VLAN configurations use 192.168.100.1/21 and 192.168.200.1/21 as VLAN prefixes. Such prefixes conflict with the BGP routes (e.g., 192.168.200.0/25). It follows that part of the VLAN addresses may be overridden by BGP routes and become unreachable. This PR aims to fix the issue by avoiding overlap between VLAN prefixes and BGP routes.

How did you do it?
Update VLAN prefixed to avoid conflict with BGP routes

How did you verify/test it?
Verify that VLAN addresses are not overlapped with BGP routes and are reachable.