Merged PR 44: [azp]: Add 'dualtor' to pytest topos for dualtors
Currently dual ToR traffic tests are being skipped during nightly tests because 'dualtor' is not being passed to the topology arg (-t) when the nightly test is run.

To fix this, check when generating nightly test params if the testbed has 'dualtor' in the name. If so, add 'dualtor' to the argument.