Changing the reboot uptime checker to be more robust
The checker is currently using the arbitrary value of 10 seconds gap between the 'now time before rebooting' and the uptime after rebooting.
However,in case of soft boot- it is almost immediate (it executes /sbin/reboot) and with powerful CPUs we see it coming up in about 8-9 seconds which fails the check below. Hence this arbitrary wait was changed and not the test only verifies that uptime is bigger than 'now time before rebooting'.