Merged PR 55: Make the merge job formal
Make the merge job formal

The merge job uses yaml file .azure-pipelines/repo_mgmt/merge_sonic_mgm.yml. This yaml file exists
in both internal and azp-test branches. The yaml file of internal branch merges to 'internal-test' branch
and builds 'internal' branch. The yaml file of azp-test branch merges to 'internal' branch and builds
azp-test branch. The merge job is configured to use the yaml file of azp-test branch. Somehow,
Azure Pipeline tried to run the job for both internal and azp-test branch. When AZP run the job for
internal branch, it uses the yaml file from the 'internal' branch. This is so confusing!

This fix is changed the yaml file of internal branch to merge from github master to 'internal' branch
too. Then we can deprecate the azp-test branch and update the job to use the yaml file from
internal branch.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>