Increase timeout for sanity check (#3826)
What is the motivation for this PR?
On multi-asic VS testbed, due to the increased number of services running, we see sanity check not able to complete critical process check before the timeout. This could be a performance issue, but increasing the timeout should help ensure that sanity check is able to finish execution.

How did you do it?
Increase the timeout to wait for completion of check_processes_on_dut parallel runs.