Directly using AZP parameter syntax in bash scripts to avoid quote issues
Currently some Azure Pipeline parameters are passed to bash scripts by environment variables.
The quotes in parameter values may not be properly processed and could cause issues. An
example is the TESTBED_SPECIFIC parameter. If we assign it value like below:

-S "platform_tests dualtor ecmp decap"

Eventually in the bash script for running tests, it is expanded as below:

-S "'platform_tests' dualtor ecmp 'decap'"

To avoid such issue, this PR removed the extra step of assigning parameter value to environment
variable. Instead, we can directly use the Azure Pipeline syntax in the bash script to get
parameter value. Then parameter value in the above example will be expanded as below:

-S 'platform_tests dualtor ecmp decap'

In this PR, not just the TESTBED_SPECIFIC parameter is updated. Other updated parameters:
* TESTBED_NAME
* TESTBED_FILE
* ALWAYS_INSTALL_NEW_IMAGE
* UPGRADE_TYPE
* PAUSE_TIME
* SKIP_SCRIPTS

Signed-off-by: Xin Wang <xiwang5@microsoft.com>