Merged PR 65: Set empty url for saithrift in RDMA nightly template
The common nightly test template was updated recently to include saithrift url as one of the extra params to "run_test.sh". This caused RDMA nightly to fail at pretest. Since this is not needed on tgen setup, passing empty url in rdma template.

Signed-off-by: Neetha John <nejo@microsoft.com>

Kicked off a pipeline run and it has passed pretest with the change
https://dev.azure.com/mssonic/internal/_build/results?buildId=27481&view=results