[chassis] Skip setupDutConfig module for modular chassis (#3788)
What is the motivation for this PR?
Teat case 'arp/test_neighbor_mac_noptf.py' brings all bgp down, In case of Modular chassis testcase fails because VOQ chassis has extra routes so when script brings bgp down and verify for the routes and expects routes should be zero never works.

How did you do it?
To Take care of this failure, we do not shutdown all BGP neighbors for a modular chassis.

The reasons are:-

When we bring BGP down using 'sudo config bgp shutdown all' on VOQ chassis linecard, it only brings down the eBGP neighbors, and not the BGP_VOQ_CHASSIS_NEIGHBORs (the iBGP neighbors to other asics in the chassis)
The asic has routes that are learnt from other remote asics in the chassis.
VoQ architecture adds static routes for inband interfaces and all eBGP peers on the remote asics as well.
To get all the routes to be flushed could be very complex. Also, the reason that the BGP shutdown was added was because on some DUT's are overwhelmed with the BGP updates and thus causing this test to intermittently fail. However, we don't see such intermittent failure on linecards of a VoQ chassis which has 12K routes in a T2 topology - possibly because CPU is powerful.