tacacs test: fixed wrong assertion which caused tests to pass even when operation failed (#4017)
TACACS tests execute a CLI command on DUT and verify its output is as expected. Command output check is implemented in the way it can be skipped when the command fails and tests pass then. Addressed the issue.
- dedicated a function for repeated test assertion fragment

Problematic code fragment:

```
res = ssh_remote_run(dut, cmd)
for l in res['stdout_lines']: # stdout_lines is empty when cmd fails so assert is never done
  assert ...
```