[test_warm_reboot_sad_bgp] add "OpenSent" as one of the expected neighbor state (#3882)
Sometimes test_warm_reboot_sad_bgp is failed because BGP neighbor state is not down in preboot state, it is due to the BGP neighbor state is "OpenSent" not "Active". VM still sends OPEN message after DUT BGP shutdown, and there is a chance to get the BGP state is "OpenSent".